+++ Tests/Unit/Csv/CsvResponseTest.php
@@ -1,0 +1,147 @@
+<?php
+
+declare(strict_types=1);
+
+namespace OliverKlee\Seminars\Tests\Unit\Csv;
+
+use Nimut\TestingFramework\TestCase\UnitTestCase;
+use OliverKlee\Oelib\Configuration\ConfigurationRegistry;
+use OliverKlee\Oelib\Configuration\DummyConfiguration;
+use OliverKlee\Seminars\Csv\CsvResponse;
+use Psr\Http\Message\ResponseInterface;
+
+/**
+ * @covers \OliverKlee\Seminars\Csv\CsvResponse
+ */
+final class CsvResponseTest extends UnitTestCase
+{
+    /**
+     * @var DummyConfiguration
+     */
+    private $configuration;
+
+    protected function setUp(): void
+    {
+        $this->configuration = new DummyConfiguration();
+        ConfigurationRegistry::getInstance()->set('plugin.tx_seminars', $this->configuration);
+    }
+
+    protected function tearDown(): void
+    {
+        ConfigurationRegistry::purgeInstance();
+
+        parent::tearDown();
+    }
+
+    /**
+     * @test
+     */
+    public function implementsResponse(): void
+    {
+        self::assertInstanceOf(ResponseInterface::class, new CsvResponse(''));
+    }
+
+    /**
+     * @test
+     */
+    public function hasCsvContentTypeWithHeader(): void
+    {
+        $subject = new CsvResponse('');
+
+        $contentTypeHeader = $subject->getHeader('Content-Type')[0];
+        self::assertStringContainsString('Content-Type: text/csv; header=present;', $contentTypeHeader);
+    }
+
+    /**
+     * @test
+     */
+    public function hasCsvContentTypeWithHeaderForLowercasedHeaderName(): void
+    {
+        $subject = new CsvResponse('');
+
+        $contentTypeHeader = $subject->getHeader('content-type')[0];
+        self::assertContains('Content-Type: text/csv; header=present; ', $contentTypeHeader);
+    }
+
+    /**
+     * @return array<string, array<int, non-empty-string>>
+     */
+    public function charsetDataProvider(): array
+    {
+        return [
+            'iso-8859-15' => ['iso-8859-15'],
+            'utf-8' => ['utf-8'],
+        ];
+    }
+
+    /**
+     * @test
+     *
+     * @dataProvider charsetDataProvider
+     */
+    public function usesCharsetFromConfigurationForTheContentType(string $charset): void
+    {
+        $this->configuration->setAsString('charsetForCsv', $charset);
+        $subject = new CsvResponse('');
+
+        $contentTypeHeader = $subject->getHeader('Content-Type')[0];
+        self::assertStringContainsString('charset=' . $charset, $contentTypeHeader);
+    }
+
+    /**
+     * @test
+     */
+    public function hasContentDispositionAttachment(): void
+    {
+        $subject = new CsvResponse('');
+
+        $contentDispositionHeader = $subject->getHeader('Content-Disposition')[0];
+        self::assertContains('Content-Disposition: attachment', $contentDispositionHeader);
+    }
+
+    /**
+     * @test
+     */
+    public function hasContentDispositionAttachmentWithLowercasedHeaderName(): void
+    {
+        $subject = new CsvResponse('');
+
+        $contentDispositionHeader = $subject->getHeader('content-disposition')[0];
+        self::assertContains('Content-Disposition: attachment', $contentDispositionHeader);
+    }
+
+    /**
+     * @test
+     */
+    public function usesProvidedFileNameInContentDisposition(): void
+    {
+        $filename = 'registrations.csv';
+        $subject = new CsvResponse('php://temp', $filename);
+
+        $contentDispositionHeader = $subject->getHeader('Content-Disposition')[0];
+        self::assertContains('; filename=' . $filename, $contentDispositionHeader);
+    }
+
+    /**
+     * @test
+     */
+    public function withoutFilenameNotUsesAnyFileNameInContentDisposition(): void
+    {
+        $subject = new CsvResponse('');
+
+        $contentDispositionHeader = $subject->getHeader('Content-Disposition')[0];
+        self::assertNotContains('filename=', $contentDispositionHeader);
+    }
+
+    /**
+     * @test
+     */
+    public function bodyHasProvidedContent(): void
+    {
+        $bodyContent = "a;b;c\r\n1;2;3";
+
+        $subject = new CsvResponse($bodyContent);
+
+        self::assertSame($bodyContent, (string)$subject->getBody());
+    }
+}